+++ src/lib/constants.ts
@@ -1,29 +1,26 @@
 import { createCanvas } from "canvas";
-import type { ChartConfiguration, ChartDataset } from "chart.js";
-import adapter from "date-fns/locale/en-US/index.js";
-import dayjs from "dayjs";
-import type { APIEmbed } from "discord.js";
-import { ToptatsExampleEmbed, UserStatsExampleEmbed } from "../types/index.js";
-import { codeString, placementSuffix } from "./helpers.js";
+import * as deepl from "deepl-node";
 
-export const statusRoles = [
+export const GLOBAL_CANVAS = createCanvas(1200, 400);
+export const CHARTJS_NODE_CANVAS = GLOBAL_CANVAS.getContext("2d");
+export const TRANSLATOR = new deepl.Translator(process.env.DEEPL!);
+
+export const STATUS_ROLES = [
   "Verified",
   "VoiceOnly",
   "Jail",
   "Unverified",
 ] as const;
 
-export const memberRoles = ["Admin", "Owner", "Helper", "Member"].map((e) =>
-  e.toLowerCase(),
-);
+export const MEMBER_ROLES = ["Admin", "Owner", "Helper", "Member"] as const;
 
 export const GENERAL_CHANNEL = "💬│𝖦𝖾𝗇𝖾𝗋𝖺𝗅";
 export const EVERYONE = "@everyone";
 export const BUMPER = "Bumper";
-export const VERIFIED = statusRoles[0];
-export const VOICE_ONLY = statusRoles[1];
-export const READ_ONLY = statusRoles[2];
-export const MUTE = statusRoles[3];
+export const VERIFIED = STATUS_ROLES[0];
+export const VOICE_ONLY = STATUS_ROLES[1];
+export const READ_ONLY = STATUS_ROLES[2];
+export const MUTE = STATUS_ROLES[3];
 
 export const VERIFY_CHANNEL = "✅│𝖵𝖾𝗋𝗂𝖿𝗒";
 export const BOT_CHANNEL = "🤖│𝖡𝗈𝗍";
@@ -39,288 +36,3 @@ export const VERIFY_TEMPLATE = "verify yourself";
 export const RED_COLOR = 0xff0000;
 export const BOT_ICON =
   "https://raw.githubusercontent.com/Don-Cryptus/coding.global-web/main/public/favicon/favicon-96x96.png";
-
-export const simpleEmbedExample: APIEmbed = {
-  color: RED_COLOR,
-  description: `*`,
-  timestamp: new Date().toISOString(),
-  footer: {
-    text: "*",
-    icon_url: BOT_ICON,
-  },
-};
-
-export const topStatsExampleEmbed = ({
-  mostActiveMessageUsers,
-  mostActiveMessageChannels,
-  mostActiveVoiceUsers,
-  mostActiveVoiceChannels,
-}: ToptatsExampleEmbed): APIEmbed => {
-  const mostActiveVoiceUserSumString = codeString(
-    mostActiveVoiceUsers
-      .reduce((acc, { sum }) => acc + sum, 0)
-      .toLocaleString("en") + " hours",
-  );
-
-  const mostActiveVoiceUsersString = mostActiveVoiceUsers?.map(
-    ({ memberId, sum, username }, i) =>
-      `${codeString(
-        placementSuffix(i + 1),
-      )} <@${memberId}>: (${username}) ${codeString(
-        sum.toLocaleString("en") + " hours",
-      )}`,
-  );
-
-  const mostActiveVoiceChannelSumString = codeString(
-    mostActiveVoiceChannels
-      .reduce((a, b) => a + b.sum, 0)
-      .toLocaleString("en") + " hours",
-  );
-
-  const mostActiveVoiceChannelString = mostActiveVoiceChannels.map(
-    ({ sum, channelId }, i) =>
-      `${codeString(placementSuffix(i + 1))} <#${channelId}>: ${codeString(
-        sum.toLocaleString("en") + " hours",
-      )}`,
-  );
-
-  const mostActiveMessageUsersSumString = codeString(
-    mostActiveMessageUsers
-      .reduce((a, b) => a + Number(b.count), 0)
-      .toLocaleString("en") + " messages",
-  );
-
-  const mostActiveMessageUsersString = mostActiveMessageUsers.map(
-    ({ count, memberId, username }, i) =>
-      `${codeString(
-        placementSuffix(i + 1),
-      )} <@${memberId}>: (${username}) ${codeString(
-        count.toLocaleString("en") + " messages",
-      )}`,
-  );
-
-  const mostActiveMessageChannelSumString = codeString(
-    mostActiveMessageChannels
-      .reduce((a, b) => a + Number(b.count), 0)
-      .toLocaleString("en") + " messages",
-  );
-
-  const mostActiveMessageChannelString = mostActiveMessageChannels.map(
-    ({ count, channelId }, i) =>
-      `${codeString(placementSuffix(i + 1))} <#${channelId}>: ${codeString(
-        Number(count).toLocaleString("en") + " messages",
-      )}`,
-  );
-
-  return {
-    color: RED_COLOR,
-    title: `⭐ Top Stats Overview`,
-
-    description: `
-Top users and channels in the past __9,999 Days__.
-
-**Messages | Top ${mostActiveMessageUsers?.length ?? 0} - Members**
-Top Member Sum: ${mostActiveMessageUsersSumString} 
-
-${mostActiveMessageUsersString.join("\n")}
-
-Messages | Top ${mostActiveMessageChannels?.length ?? 0} - Channels
-Top Channel Sum: ${mostActiveMessageChannelSumString}
-
-${mostActiveMessageChannelString.join("\n")}
-
-**Voice | Top ${mostActiveVoiceUsers?.length ?? 0} - Members**
-Top Member Sum: ${mostActiveVoiceUserSumString} 
-
-${mostActiveVoiceUsersString.join("\n")}
-
-Voice | Top ${mostActiveVoiceChannels?.length ?? 0} - Channels
-Top Channel Sum: ${mostActiveVoiceChannelSumString}
-
-${mostActiveVoiceChannelString.join("\n")}
-
-`,
-    timestamp: new Date().toISOString(),
-    footer: {
-      text: TOP_STATS_TEMPLATE,
-      icon_url: BOT_ICON,
-    },
-  };
-};
-
-export const userStatsExampleEmbed = ({
-  id,
-  userGlobalName,
-  userServerName,
-  lookback,
-  joinedAt,
-  createdAt,
-  lookbackDaysCount,
-  sevenDaysCount,
-  oneDayCount,
-  mostActiveTextChannelId,
-  mostActiveTextChannelMessageCount,
-  lastMessage,
-  lastVoice,
-  mostActiveVoice,
-  lookbackVoiceSum,
-  oneDayVoiceSum,
-  sevenDayVoiceSum,
-}: UserStatsExampleEmbed): APIEmbed => {
-  const lookbackCommand = codeString("/lookback-me");
-  const mostActiveTextChannelString = mostActiveTextChannelId
-    ? `<#${mostActiveTextChannelId}>`
-    : "";
-  const mostActiveTextChannelCountString = codeString(
-    mostActiveTextChannelMessageCount.toLocaleString("en") + " messages",
-  );
-
-  const mostActiveVoiceChannelString = mostActiveVoice?.channelId
-    ? `<#${mostActiveVoice.channelId}> ${codeString(
-        mostActiveVoice?.sum.toLocaleString("en") + " hours",
-      )}`
-    : "";
-
-  const lookbackVoiceSumString = codeString(
-    lookbackVoiceSum.toLocaleString("en") + " hours",
-  );
-  const oneDayVoiceSumString = codeString(
-    oneDayVoiceSum.toLocaleString("en") + " hours",
-  );
-  const sevenDaysVoiceSumString = codeString(
-    sevenDayVoiceSum.toLocaleString("en") + " hours",
-  );
-
-  const joinedAtUnix = dayjs(joinedAt).unix();
-  const createdAtUnix = dayjs(createdAt).unix();
-  const lastMessageUnix =
-    lastMessage?.length > 0 && dayjs(lastMessage[0]!.createdAt).unix();
-  const lastVoiceUnix =
-    lastVoice?.length > 0 && dayjs(lastVoice?.[0]?.leave ?? new Date()).unix();
-  const lastMessageString = lastMessageUnix
-    ? `__<t:${lastMessageUnix}:D>__ (<t:${lastMessageUnix}:R>)`
-    : codeString("None");
-
-  const lastVoiceString = lastVoiceUnix
-    ? `__<t:${lastVoiceUnix}:D>__ (<t:${lastVoiceUnix}:R>)`
-    : codeString("None");
-
-  return {
-    color: RED_COLOR,
-    title: `👤 ${userGlobalName}'s Stats Overview`,
-
-    description: `
-${userServerName} (${userGlobalName})
-
-User stats in the past __${lookback}__ Days. (Change with the ${lookbackCommand} command.)
-
-**User Info**
-Joined On: __<t:${joinedAtUnix}:D>__ (<t:${joinedAtUnix}:R>)
-Created On: __<t:${createdAtUnix}:D>__ (<t:${createdAtUnix}:R>)
-Last Message On: ${lastMessageString}
-Last Voice On: ${lastVoiceString}
-User ID: ${codeString(id)}
-
-**Most Active Channels**
-Messages: ${mostActiveTextChannelString} ${mostActiveTextChannelCountString}
-Voice: ${mostActiveVoiceChannelString}
-`,
-
-    fields: [
-      {
-        name: "Messages",
-        value: `
-__${lookback} Days__: ${codeString(
-          `${lookbackDaysCount.toLocaleString("en")} messages`,
-        )} 
-7 Days: ${codeString(`${sevenDaysCount.toLocaleString("en")} messages`)}
-24 Hours: ${codeString(`${oneDayCount.toLocaleString("en")} messages`)}
-`,
-
-        inline: true,
-      },
-      {
-        name: "Voice",
-        value: `
-__${lookback} Days__: ${lookbackVoiceSumString}
-7 Days: ${sevenDaysVoiceSumString}
-24 Hours: ${oneDayVoiceSumString}
-`,
-        inline: true,
-      },
-    ],
-
-    timestamp: new Date().toISOString(),
-    footer: {
-      text: STATS_TEMPLATE,
-      icon_url: BOT_ICON,
-    },
-  };
-};
-
-export const chartConfig = (data: ChartDataset[]) => {
-  return {
-    type: "line",
-    data: {
-      datasets: [
-        {
-          data,
-          fill: true,
-          backgroundColor: "#495170",
-          borderColor: "#6f86d4",
-        },
-      ],
-    },
-    plugins: [
-      {
-        id: "customCanvasBackgroundColor",
-        beforeDraw: (chart, _args, options) => {
-          const { ctx } = chart;
-          ctx.save();
-          ctx.globalCompositeOperation = "destination-over";
-          ctx.fillStyle = options.color || "#34363c";
-          ctx.fillRect(0, 0, chart.width, chart.height);
-          ctx.restore();
-        },
-      },
-    ],
-    options: {
-      showLine: true,
-      elements: {
-        point: {
-          radius: 0,
-        },
-      },
-      plugins: {
-        legend: {
-          display: false,
-        },
-        customCanvasBackgroundColor: {
-          color: "#34363c",
-        },
-      },
-
-      scales: {
-        yAxes: {
-          ticks: { color: "#fff" },
-          grid: { color: "#e6e6e6", z: 100, drawBorder: false },
-          // max: data.length + Math.round((data.length / 100) * 25),
-        },
-        xAxes: {
-          ticks: { color: "#fff" },
-          grid: { display: false, drawBorder: false },
-          type: "timeseries",
-          adapters: {
-            date: {
-              locale: adapter,
-            },
-          },
-          time: { unit: data.length > 360 ? "month" : "day" },
-        },
-      },
-    },
-  } as ChartConfiguration<"line", ChartDataset[]>;
-};
-export const globalCanvas = createCanvas(1200, 400);
-
-export const chartJSNodeCanvas = globalCanvas.getContext("2d");
