--- src/types/discord.ts
@@ -1,7 +1,0 @@
-import type { Collection } from 'discord.js';
-
-declare module 'discord.js' {
-  export interface Client {
-    commands: Collection<unknown, any>;
-  }
-}
